refactor: migrate FormComponent and FormComponentType imports to FormBuilder.types
